+++ packages/core/src/components/axes/zoom-bar.ts
@@ -10,7 +10,6 @@ import {
 } from "../../interfaces";
 import { DOMUtils } from "../../services";
 import * as Configuration from "../../configuration";
-import isEmpty from "lodash/isEmpty";
 
 // D3 Imports
 import { extent } from "d3-array";
@@ -129,7 +128,7 @@ export class ZoomBar extends Component {
 
 		if (mainXScale && mainXScaleType === ScaleTypes.TIME) {
 			let zoomBarData = this.services.zoom.getZoomBarData();
-			if (isEmpty(zoomBarData)) {
+			if (Tools.isEmpty(zoomBarData)) {
 				// if there's no zoom bar data we can't do anything
 				return;
 			}