+++ packages/core/src/model.ts
@@ -4,8 +4,7 @@ import { Tools } from "./tools";
 import {
 	Events,
 	ScaleTypes,
-	ColorClassNameTypes,
-	AxisPositions
+	ColorClassNameTypes
 } from "./interfaces";
 
 // D3
@@ -577,44 +576,8 @@ export class ChartModel {
 		return data;
 	}
 
-	protected sanitizeDateValues(data) {
-		const options = this.getOptions();
-
-		if (!options.axes) {
-			return data;
-		}
-
-		const keysToCheck = [];
-		Object.keys(AxisPositions).forEach((axisPositionKey) => {
-			const axisPosition = AxisPositions[axisPositionKey];
-			const axisOptions = options.axes[axisPosition];
-
-			if (axisOptions && axisOptions.scaleType === ScaleTypes.TIME) {
-				const axisMapsTo = axisOptions.mapsTo;
-
-				if (axisMapsTo !== null || axisMapsTo !== undefined) {
-					keysToCheck.push(axisMapsTo);
-				}
-			}
-		});
-
-		if (keysToCheck.length > 0) {
-			// Check all datapoints and sanitize dates
-			data.forEach((datum) => {
-				keysToCheck.forEach((key) => {
-					if (datum[key].getTime === undefined) {
-						datum[key] = new Date(datum[key]);
-					}
-				});
-			});
-		}
-
-		return data;
-	}
-
 	protected sanitize(data) {
 		data = this.getTabularData(data);
-		data = this.sanitizeDateValues(data);
 
 		return data;
 	}
@@ -688,6 +651,12 @@ export class ChartModel {
 
 		const options = this.getOptions();
 		const userProvidedScale = Tools.getProperty(options, "color", "scale");
+
+		Object.keys(userProvidedScale).forEach(dataGroup => {
+			if (!this.allDataGroups.includes(dataGroup)) {
+				console.warn(`"${dataGroup}" does not exist in data groups.`);
+			}
+		})
 
 		/**
 		 * Go through allDataGroups. If a data group has a color value provided